# Optimised find_center_vo_cupy