This optimises the function `find_center_vo_cupy` for the GPU, as far as it goes without changing the algorithm / logic involved. 

### Performance Improvement: 255ms -> 194ms (1.31x)

(This was measured on V100 using the performance test that is part of this PR.)

### Remaining Bottlenecks

The key bottleneck however are the spline shifts for sub-pixel shifting - they dominate the overall time by far (ignoring the FFT), especially after these optimisations. We get the following picture for the kernel percentages:

| Function | Percent GPU before | Percent GPU after |
|---|---:|---:|
| FFT | 44.3% | 46.4% |
| Spline Shift | 29.8% | 33.6% |
| all other kernels combined | 20.4% | 19.2% |
| Memory operations | 5.5% | 0.8% |

(Note that the FFT dimensions are not powers of 2 - this causes the long FFT time. Also note that a big part of the optimisation was reducing the number of kernel calls and hereby reducing gaps where the GPU is idle.)

### Details of Changes

- small operations such as ceil, div, abs, of scalars are computed on the CPU instead of launching kernels
- other small operations where joined on the GPU
- axis extensions and copies of data are avoided
- a single RawKernel does:
   - integer shifts of the input data for all lines where the shift is integer
   - inserts the `sino3` array into the part which was shifted out
 - then the scipy ndimage shift function (using order 3 splines with prefiltering) is only called for the rows fractional shifts, in a loop
 - the FFT, masking, and mean calculation is done once for all possible shifts (outside of the loop)
 - the downsample and mask generation kernels have been optimised further